Mavenify directory structure; many updates.
The directory structure is now "src/main/java" and "src/test/java".
Numerous model classes were also added (TrainingActivity, TrainingWeek,
TrainingPlan, MarathonType, Weekday). A French language resource was
added. TrainerFileLoader was inserted (as well as tests for it) - it's
responsible for loading a training plan given a MarathonType and a path.